requirements.txt
# Cyberdeck Code Requirements
# Install with: pip install -r requirements.txt

# GPIO Control for Raspberry Pi 5
gpiozero>=2.0.1

# Servo Control via PCA9685
adafruit-circuitpython-servokit>=1.3.21
adafruit-circuitpython-pca9685>=3.4.18

# OLED Display Control
luma.oled>=3.10.0
luma.core>=2.4.0
Pillow>=9.0.0
smbus2>=0.4.2

# Optional but recommended
numpy>=1.21.0  # For advanced animations